patch from From: Jim Hayward <jimhayward@linuxexperience.com>
This patch adds keyboard mnemonics to all of the dialogs that did not
have them before. Quite a few of the dialogs still need some layout
redesigns I think. For now just the keyboard mnemonics and some minor
changes to a couple of the dialogs.

also some changes by Olivier to get the php icon back in place


git-svn-id: svn://svn.code.sf.net/p/bluefish/code/trunk/bluefish@396 7640d186-872a-4c71-b7b2-a5696852cc35