from Michele Garoche <michele.garoche@easyconnect.fr>
To make it work with the new names of custom_menu.xx.default.

I've also changed LANG to LANGUAGE (it does not work with LANG on Mac,
I don't know if it is specific to Mac or not), in case it is specific
I'll merely patch locally to have LANGUAGE instead LANG in case of
reverting to LANG.

I've also added three printf just to be sure the correct file is loaded.


git-svn-id: svn://svn.code.sf.net/p/bluefish/code/trunk/bluefish@1380 7640d186-872a-4c71-b7b2-a5696852cc35