From: Timothy Joko-Veltman <tim-oss@indigopixel.com>
Here is a new highlighting file.  Based on today's CVS.

Fixed:
   XML highlighting parentmatching rules
   Minor CSS rule errors
   PHP highlighting parentmatching rules
Added
  C Functions, Pointers, Variables rules

Notes:
I also reviewed other filetypes for parentmatching problems, and fixed
them as appropriate.

Colors are the defaults, except in the C rules I added; tried to choose
pleasing and accessible colors.

I couldn't find Tcl rules ... There was a spot in the menu for them, but
no rules loaded when selected.  The same is true for object, image and
webimage filetypes.


git-svn-id: svn://svn.code.sf.net/p/bluefish/code/trunk/bluefish@1383 7640d186-872a-4c71-b7b2-a5696852cc35