I think I fixed the issue reported by Jim: The segfault is at line 572. &tmpiter is not valid when it reaches this point. The segfault happens most of the time right after the last item in the tree is deleted.
In bmark_del_children_backend() the pointer "parent" is still a valid
pointer, but it is not a valid iter. It's already been deleted in the
call to bmark_check_remove().

The call at line 570 to gtk_tree_model_iter_children() appears to return
TRUE even though all items in the tree have been deleted. The number of
children also still returns as 1 on "parent" even though this is not a
valid iter.

Explanation is a little simplified. I'm sure I forgot something too. I
need to get up from the computer and go to work.  :-P


git-svn-id: svn://svn.code.sf.net/p/bluefish/code/trunk/bluefish@3179 7640d186-872a-4c71-b7b2-a5696852cc35