Added a check and and a variable for gdk-pixbuf-csource just for the case, it is not in PATH.
Fixed the --disable|enable-libgnomeui configure option description.

Fixed a small typo.

Moved the tools determination out of AC_ARG_ENABLE for update-databases.
Added variables to set the tools location for the case, they are not in
PATH. Further removed AC_SUBST() for these variables, because AC_CHECK_PROG
calls it.


git-svn-id: svn://svn.code.sf.net/p/bluefish/code/trunk@3564 7640d186-872a-4c71-b7b2-a5696852cc35