Disabled the build of statically linked plugins by default.
Added a comment for later. AM_MAINTAINER_MODE can turn off automatically
triggering an autotool-chain re-run if a marcro-file like configure.ac or
Makefile.am changes. This is probably better for packages/distributors.
The rules to automatically re-run aclocal/autoheader/autoconf/automake
can be re-enabled with the --enable-maintainer-mode option in this case.


git-svn-id: svn://svn.code.sf.net/p/bluefish/code/trunk/bluefish@3600 7640d186-872a-4c71-b7b2-a5696852cc35