fixed http://bugzilla.gnome.org/show_bug.cgi?id=566537
When I'm writing in PHP any variables I place in text strings are highlighted
as text strings, not variables.  It would be a nice help to have all variables,
inside and outside strings coloured the same way.  This is especially useful
when writing myswl commands in PHP as they need to be writing inside a text
string.



git-svn-id: svn://svn.code.sf.net/p/bluefish/code/trunk@4410 7640d186-872a-4c71-b7b2-a5696852cc35