fix the bug reported by Roberto Bellingeri
The bug was in the scancache-offset-update code, only triggered by multiple replaces (because a replace 
does two offset updates on the same location). Improved the code such that scancache_update_single_offset()
now runs only once for such a situation (and handles both the insert and the delete simultaneous), by
adding an extra parameter 'deleteoffset' that indicates which entries in the cache can be removed.



git-svn-id: svn://svn.code.sf.net/p/bluefish/code/trunk/bluefish@8376 7640d186-872a-4c71-b7b2-a5696852cc35